extract inner classes for ListCells out of ApplicationController
